adding notes
